The installer has no device unless user asks to view disk layout
The dialog command showing available disks expects to have them in the 'disks'
array. The disks are read by

   diskinfo="$(get_disks)"

But since get_disks is executed in a subshell here, the varialbe 'disks' does
not propagate to the main shell. To fix that I am executing the function
get_disks directly in the main shell. The function used to also print to stdout
parameters for the dialog command. But the functionality seems to be common
enough to make extra function which will take the array itself as an input. It
also has the advantage of handling spaces correctly compared to passing the
arguments as a string.